# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lib API",
        "description": "API for managing books",
        "version": "1.0"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "responses": {
                    "200": {
                        "description": "A list of books"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "operationId": "47047e95a700b3271b697e57d0965191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a book by ID",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book by ID",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book by ID",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/books/search": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Search for books by query",
                "operationId": "a50e12f6ebf00d1971a30d2fe743509c",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Author name or book title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of books matching the search criteria"
                    }
                }
            }
        },
        "/api/books/{id}/reserve": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Reserve a book by ID",
                "operationId": "48b9babc0d90aa559d70dc5a182facbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Book reserved"
                    },
                    "400": {
                        "description": "Book is already reserved"
                    },
                    "401": {
                        "description": "Token not provided"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/books/{id}/return": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Return a reserved book by ID",
                "operationId": "3ebaaf1ff1d3c8c3fc4ac6082f3151d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned"
                    },
                    "400": {
                        "description": "Book is not reserved"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get all reservations",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "responses": {
                    "200": {
                        "description": "A list of reservations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Create a new reservation",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reservation"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created"
                    },
                    "400": {
                        "description": "Book already reserved"
                    },
                    "404": {
                        "description": "Book or User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get a reservation by ID",
                "operationId": "a226b43b2a6a846245e658d01bf00720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation details"
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Delete a reservation by ID",
                "operationId": "9b6641e4ded1ff7adfce352d589b85e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reservation deleted"
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "500": {
                        "description": "Serve error"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register user",
                "operationId": "d03ee3a44bdd6f0025efd14c5c76b5ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User create successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User create successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "required": [
                    "title",
                    "author",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Idiot"
                    },
                    "author": {
                        "type": "string",
                        "example": "F.M. Dostoevsky"
                    },
                    "release_date": {
                        "type": "integer",
                        "example": "1896-01-01"
                    }
                },
                "type": "object"
            },
            "Reservation": {
                "required": [
                    "book_id",
                    "user_id"
                ],
                "properties": {
                    "book_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Alesha"
                    },
                    "email": {
                        "type": "string",
                        "example": "alesha12@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "supersecretpassword123"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "token_security": {
                "type": "apiKey",
                "description": "Your api key",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Operations with books"
        },
        {
            "name": "Reservation",
            "description": "Operations with reservation"
        },
        {
            "name": "User",
            "description": "Operations with user"
        }
    ],
    "security": [
        {
            "token_security": {
                "type": "apiKey",
                "name": "token",
                "in": "header"
            }
        }
    ]
}